Extract init functions of supported physical components to own files
