+++ services/physical/lib/factories/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  Button: require('./Button'),
+  Capacitive: require('./Capacitive'),
+  Encoder: require('./Encoder'),
+  'Led.RGB': require('./Led.RGB')
+};
